Recommend setting OPTLPENDING to the empty string before first call
If the previous run of getopts_long was interrupted before processing
the full list of options, then OPTLPENDING may still contain some
pending options which would affect the new run for a different set of
options.

So, in the example and the test code, we set it to the empty string
before the first run. It's also good practice to initialise variables
before their first use.

Note that as noted at
https://www.mail-archive.com/austin-group-l@opengroup.org/msg04114.html
that also affects some implementations of getopts and those don't expose
their "pending" list to the user, so there's no easy way to reset it
there.